refactor: split react.test.tsx in multiple files (#2207)
* refactor: split react.test.tsx in multiple files

* refactor: rename createAppRouter to createLegacyAppRouter

This will make it more clear that these are interop tests that are using the
old way (v9) of creating a tRPC router.
