+++ packages/server/test/interop/react/setQueryData.test.tsx
@@ -1,0 +1,95 @@
+/* eslint-disable @typescript-eslint/no-empty-function */
+import { createLegacyAppRouter } from './__testHelpers';
+import '@testing-library/jest-dom';
+import { render, waitFor } from '@testing-library/react';
+import React, { useState } from 'react';
+import { QueryClient, QueryClientProvider, setLogger } from 'react-query';
+
+setLogger({
+  log() {},
+  warn() {},
+  error() {},
+});
+
+let factory: ReturnType<typeof createLegacyAppRouter>;
+beforeEach(() => {
+  factory = createLegacyAppRouter();
+});
+afterEach(() => {
+  factory.close();
+});
+
+describe('setQueryData()', () => {
+  test('without & without callback', async () => {
+    const { trpc, client } = factory;
+    function MyComponent() {
+      const utils = trpc.useContext();
+      const allPostsQuery = trpc.useQuery(['allPosts'], {
+        enabled: false,
+      });
+      const postByIdQuery = trpc.useQuery(['postById', '1'], {
+        enabled: false,
+      });
+      return (
+        <>
+          <pre>{JSON.stringify(allPostsQuery.data ?? null, null, 4)}</pre>
+          <pre>{JSON.stringify(postByIdQuery.data ?? null, null, 4)}</pre>
+          <button
+            data-testid="setQueryData"
+            onClick={async () => {
+              utils.setQueryData(
+                ['allPosts'],
+                [
+                  {
+                    id: 'id',
+                    title: 'allPost.title',
+                    createdAt: Date.now(),
+                  },
+                ],
+              );
+              const newPost = {
+                id: 'id',
+                title: 'postById.tmp.title',
+                createdAt: Date.now(),
+              };
+              utils.setQueryData(['postById', '1'], (data) => {
+                expect(data).toBe(undefined);
+                return newPost;
+              });
+              // now it should be set
+              utils.setQueryData(['postById', '1'], (data) => {
+                expect(data).toEqual(newPost);
+                if (!data) {
+                  return newPost;
+                }
+                return {
+                  ...data,
+                  title: 'postById.title',
+                };
+              });
+            }}
+          />
+        </>
+      );
+    }
+    function App() {
+      const [queryClient] = useState(() => new QueryClient());
+      return (
+        <trpc.Provider {...{ queryClient, client }}>
+          <QueryClientProvider client={queryClient}>
+            <MyComponent />
+          </QueryClientProvider>
+        </trpc.Provider>
+      );
+    }
+
+    const utils = render(<App />);
+
+    utils.getByTestId('setQueryData').click();
+
+    await waitFor(() => {
+      expect(utils.container).toHaveTextContent('allPost.title');
+      expect(utils.container).toHaveTextContent('postById.title');
+    });
+  });
+});
